moved compass-components for ProductSwitcher to separate component
